Split up AppStatusBannerComponent
The outcome banner logic goes into AppOutcomeBannerComponent,
the simple banner logic goes into AppSimpleOutcomeBannerComponent
