+++ spec/components/previews/app_outcome_banner_component_preview.rb
@@ -1,0 +1,23 @@
+class AppOutcomeBannerComponentPreview < ViewComponent::Preview
+  def triaged_do_not_vaccinate
+    patient_session =
+      FactoryBot.create(:patient_session, :triaged_do_not_vaccinate)
+
+    render AppOutcomeBannerComponent.new(patient_session:)
+  end
+
+  def unable_to_vaccinate_with_contradications
+    patient_session = FactoryBot.create(:patient_session, :unable_to_vaccinate)
+
+    render AppOutcomeBannerComponent.new(patient_session:)
+  end
+
+  def vaccinated
+    patient_session = FactoryBot.create(:patient_session, :vaccinated)
+    patient_session.vaccination_records.first.update!(
+      notes: "Patient felt dizzy"
+    )
+
+    render AppOutcomeBannerComponent.new(patient_session:)
+  end
+end
